Added registration and comment section

+++ blog/views.py
@@ -4,6 +4,8 @@ from django.utils import timezone
 from .forms import PostForm, CommentForm
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth import logout
+from django.contrib.auth.models import User
+
 
 
 def post_list(request):
@@ -14,51 +16,68 @@ def post_detail(request, pk):
     post = get_object_or_404(Post, pk=pk)
     return render(request, "blog/post_detail.html", {'post': post})
 
-@login_required
+# @login_required
 def post_new(request):
-    if request.method == "POST":
-        form = PostForm(request.POST)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            # post.published_date = timezone.now()
-            post.save()
-            return redirect('post_detail', pk=post.pk)
+    if request.user.is_superuser:
+        if request.method == "POST":
+            form = PostForm(request.POST)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.author = request.user
+                # post.published_date = timezone.now()
+                post.save()
+                return redirect('post_detail', pk=post.pk)
+        else:
+            form = PostForm()
+        return render(request, 'blog/post_edit.html', {'form': form})
     else:
-        form = PostForm()
-    return render(request, 'blog/post_edit.html', {'form': form})
+        return redirect('post_list')
 
-@login_required
+# @login_required
 def post_edit(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    if request.method == "POST":
-        form = PostForm(request.POST, instance=post)
-        if form.is_valid():
-            post = form.save(commit=False)
-            post.author = request.user
-            # post.published_date = timezone.now()
-            post.save()
-        return redirect('post_detail', pk=post.pk)
+    if request.user.is_superuser:
+        if request.method == "POST":
+            form = PostForm(request.POST, instance=post)
+            if form.is_valid():
+                post = form.save(commit=False)
+                post.author = request.user
+                # post.published_date = timezone.now()
+                post.save()
+            return redirect('post_detail', pk=post.pk)
+        else:
+            form = PostForm(instance=post)
+        return render(request, 'blog/post_edit.html', {'form': form})
     else:
-        form = PostForm(instance=post)
-    return render(request, 'blog/post_edit.html', {'form': form})
+        return redirect('post_detail', pk=post.pk)
 
-@login_required
+# @login_required
 def post_draft_list(request):
-    posts = Post.objects.filter(published_date__isnull=True).order_by('-created_date')
-    return render(request, 'blog/post_draft_list.html', {'posts': posts})
+    if request.user.is_superuser:
+        posts = Post.objects.filter(published_date__isnull=True).order_by('-created_date')
+        return render(request, 'blog/post_draft_list.html', {'posts': posts})
+    else:
+        return redirect('post_list')
 
-@login_required
+# @login_required
 def post_publish(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    post.publish()
-    return redirect('post_detail', pk=pk)
+    if request.user.is_superuser:
+        post.publish()
+        return redirect('post_detail', pk=pk)
+    else:
+        return redirect('post_detail', pk=pk)
 
-@login_required
+
+# @login_required
 def post_delete(request, pk):
-    post = get_object_or_404(Post, pk=pk)
-    post.delete()
-    return redirect('post_list')
+    if request.user.is_superuser:
+        post = get_object_or_404(Post, pk=pk)
+        post.delete()
+        return redirect('post_list')
+    else:
+        return redirect('post_list')
+
 
 def post_logout(request):
     logout(request)
@@ -66,29 +85,48 @@ def post_logout(request):
 
 def add_comment_to_post(request, pk):
     post = get_object_or_404(Post, pk=pk)
-    if request.method == "POST":
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.post = post
-            comment.save()
-            return redirect('post_detail', pk=post.pk)
+    if request.user.is_authenticated():
+        name = request.user.username
+        form = CommentForm(initial={'author': name})
+        if request.method == "POST":
+            form = CommentForm(request.POST)
+            if form.is_valid():
+                comment = form.save(commit=False)
+                comment.post = post
+                comment.approve()
+                comment.save()
+                return redirect('post_detail', pk=post.pk)
     else:
-            form = CommentForm()
+        form = CommentForm()
+        if request.method == "POST":
+            form = CommentForm(request.POST)
+            if form.is_valid():
+                comment = form.save(commit=False)
+                comment.post = post
+                comment.save()
+                return redirect('post_detail', pk=post.pk)
+    # else:
+            # form = CommentForm()
     return render(request, 'blog/add_comment_to_post.html', {'form': form})
 
-@login_required
+# @login_required
 def comment_approve(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    comment.approve()
-    return redirect('post_detail', pk=comment.post.pk)
+    if request.user.is_superuser:
+        comment.approve()
+        return redirect('post_detail', pk=comment.post.pk)
+    else:
+        return redirect('post_detail', pk=comment.post.pk)
 
-@login_required
-def comment_remove(request,pk):
+# @login_required
+def comment_remove(request, pk):
     comment = get_object_or_404(Comment, pk=pk)
-    post_pk = comment.post.pk
-    comment.delete()
-    return redirect('post_detail', pk=post_pk)
+    if request.user.is_superuser:
+        post_pk = comment.post.pk
+        comment.delete()
+        return redirect('post_detail', pk=post_pk)
+    else:
+        return redirect('post_detail', pk=comment.post.pk)
